chore: extract MusicTrackTitleInput component
